// wls-gui-admintool/src/api/wls-clients/generated-basisdaten-api/api.ts
export interface AsyncProgressDTO {
    /**
     * 
     * @type {string}
     * @memberof AsyncProgressDTO
     */
    'forWahltag'?: string;
    /**
     * 
     * @type {string}
     * @memberof AsyncProgressDTO
     */
    'wahlNummer'?: string;
    /**
     * 
     * @type {string}
     * @memberof AsyncProgressDTO
     */
    'lastStartTime'?: string;
    /**
     * 
     * @type {string}
     * @memberof AsyncProgressDTO
     */
    'lastFinishTime'?: string;
    /**
     * 
     * @type {boolean}
     * @memberof AsyncProgressDTO
     */
    'wahlvorschlaegeLoadingActive'?: boolean;
    /**
     * 
     * @type {number}
     * @memberof AsyncProgressDTO
     */
    'wahlvorschlaegeTotal'?: number;
    /**
     * 
     * @type {number}
     * @memberof AsyncProgressDTO
     */
    'wahlvorschlageFinished'?: number;
    /**
     * 
     * @type {string}
     * @memberof AsyncProgressDTO
     */
    'wahlvorschlaegeNext'?: string;
    /**
     * 
     * @type {boolean}
     * @memberof AsyncProgressDTO
     */
    'referendumLoadingActive'?: boolean;
    /**
     * 
     * @type {number}
     * @memberof AsyncProgressDTO
     */
    'referendumVorlagenTotal'?: number;
    /**
     * 
     * @type {number}
     * @memberof AsyncProgressDTO
     */
    'referendumVorlagenFinished'?: number;
    /**
     * 
     * @type {string}
     * @memberof AsyncProgressDTO
     */
    'referendumVorlagenNext'?: string;
}